refactor: 인프라에서 dao를 의존하지 않도록 변경
