+++ src/main/java/wooteco/subway/domain/PathFinder.java
@@ -3,5 +3,9 @@ package wooteco.subway.domain;
 import java.util.List;
 
 public interface PathFinder {
-    Path findShortestPath(List<Section> sections, Long sourceId, Long targetId);
+    Path findShortestPathByGraph(Station source, Station target);
+
+    void addVertex(List<Station> stations);
+
+    void addEdge(Station upStation, Station downStation, Section section);
 }
